Lab_Working_1/requirements.txt
django
django-crispy-forms==1.8.1
bootstrap4==0.1.0
django-multiselectfield==0.1.10